requirements.txt
numpy
pandas
yfinance
prophet
scipy
streamlit
altair
beautifulsoup4
requests
google-generativeai